[#22] Implement first version of greedy local search
The method takes five parameters:

- `x_init`: an initial solution represented as a vector of integers.
- `n_iters`: the number of iterations to perform.
- `paths`: a matrix of distances between cities.
- `epsilon`: a small positive number used as a stopping criterion (default value is 0.001).
- `steepest`: a boolean flag indicating whether to use steepest ascent hill climbing (default value is `false`).

The method starts by initializing `x` and `x_best` with `x_init`. Here, `x` represents the current solution, and `x_best` represents the best solution found so far.

The method then enters a loop that runs for `n_iters` iterations. At each iteration, it generates a random neighbor `x` by calling the `find_random_neighbor` method and assigns it to `y`. `find_random_neighbor` picks a random neighbor from a list of randomly modified cycles. The list is created using the `generate_neighbors` function, which modifies the original cycle. The modification consists in selecting two random vertices that will be reversed and rearrange the loop according to the new order. The number of generated cycles is set during the initialization of the function.

Then, it compares the fitness of `x` and `y` using the `fitness` function. If `y` has a lower fitness (i.e., a shorter total distance in the TSP), it replaces `x` with `y`.

After that, it checks if the fitness of `x` is lower than the fitness of `x_best`. If it is, it updates `x_best` with `x`. Otherwise, if `steepest` is `true`, it resets `x` to `x_best`. This step is what differentiates steepest ascent hill climbing from simple hill climbing. In steepest ascent, the algorithm always reverts back to the best solution found so far if it can't find a better neighbor.

Finally, after the loop has finished, the method returns `x_best`, which represents the best solution found during the search.

Bibliography: https://medium.com/@nour.oulad.moussa/exploring-tsp-with-hill-climbing-algorithms-9c4762ecd483