[#24] Implemented first version of random walk
This function takes the initial cycle from which the random walk starts, the number of iterations to perform the random walk.  Inside the for loop, a random movement is selected and after which the loop is updated and delta is calculated, if delta was greater than the previous one, this cycle becomes the best one.